+++ lib/kuby/plugins/rails_app/generators/kuby.rb
@@ -109,6 +109,14 @@ class KubyGenerator < Rails::Generators::Base
     )
   end
 
+  def add_cockroach_adapter
+    if Rails::VERSION::STRING > '5.2'
+      gem 'activerecord-cockroachdb-adapter', version: '~> 6.0'
+    else
+      gem 'activerecord-cockroachdb-adapter', version: '~> 5.2'
+    end
+  end
+
   private
 
   def app_name