+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+* Modify the Rails generator to add the `activerecord-cockroachdb-adapter` to your bundle.
+
 ## 0.18.0
 * Add the ability to specify your app's root directory.
   - Call `app_root <path>` inside the `docker` section of your Kuby config file.